Refactor: dto를 response, request구분
