+++ backend/src/main/java/com/mohaeng/backend/member/dto/response/MemberLoginDto.java
@@ -1,4 +1,4 @@
-package com.mohaeng.backend.member.dto;
+package com.mohaeng.backend.member.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
